+++ app/assets/javascripts/Components/submission_table.jsx
@@ -283,7 +283,7 @@ class RawSubmissionTable extends React.Component {
   };
 
   // Submission table actions
-  collectSubmissions = (override, collect_current, apply_late_penalty) => {
+  collectSubmissions = (override, collect_current, apply_late_penalty, retain_existing_grading) => {
     this.setState({showCollectSubmissionsModal: false});
     $.post({
       url: Routes.collect_submissions_course_assignment_submissions_path(
@@ -295,6 +295,7 @@ class RawSubmissionTable extends React.Component {
         override: override,
         collect_current: collect_current,
         apply_late_penalty: apply_late_penalty,
+        retain_existing_grading: retain_existing_grading,
       },
     });
   };